Revert "post and blog"
This reverts commit 0815e95bfad6146dbdbcf2e164d37fbdfcdc50ac.
