+++ ahia.egg-info/SOURCES.txt
@@ -12,5 +12,4 @@ ahia.egg-info/dependency_links.txt
 ahia.egg-info/requires.txt
 ahia.egg-info/top_level.txt
 ahia/migrations/0001_initial.py
-ahia/migrations/0002_delete_post.py
 ahia/migrations/__init__.py
