--- blog/templates/details_of_post.html
@@ -1,47 +1,0 @@
-{% extends "blog/base.html" %}
-{% block content %}
-<div class="jumbotron mt-3 shadow-lg">
-<article class="media content-section shadow ">
-   <a href="{{ object.post_author.profile.image.url }} " target="_blank"> <img src="{{ object.post_author.profile.image.url }}  " alt="" class="article-img rounded-circle img-thumbnail img-fluid">
-   </a>
-    <div class="media-body">
-      <div class="">
-        <div class="d-flex justify-content-between">
-        <a class="mr-2 mt-0 mb-0  text-dark" href="{% url 'user-posts' object.post_author.username  %}"><h3>{{ object.post_author }}</h3></a>
-        <small class="text-muted  text-align-right">{{ object.date_upload }}</small>
-        </div>
-        {% if object.post_author != user %}
-          <div class="form-group mt-0 pt-0 d-flex justify-content-end">          
-            <a href="{{ object.post_asset.url }}" download class="btn btn-primary flex-last" type="submit"><i class="fas fa-download"></i> Downlaod</a>
-          </div>
-        {% endif %}
-        <hr>
-        {% if object.post_author == user %}
-          <div>
-            <a class="btn btn-secondary btn-sm mt-1 mb-1" href="{% url 'post-update' object.id %}">Update</a>
-            <a class="btn btn-danger btn-sm mt-1 mb-1" href="{% url 'post-delete' object.id %}">Delete</a>
-          </div>
-        {% endif %}
-      </div>
-      <div>
-        {% if object.post_asset %}
-          {% if object.extension == '.jpg' or object.extension == 'jpeg' or object.extension == 'png' %}
-            <a href="{{ object.post_asset.url }}"  download ><img src="{{ object.post_asset.url }}" alt="" class="img-fluid "></a>
-            <a href="{{ object.post_asset.url }}"  download class="text-dark"><h5>{{ object.post_asset }}</h5></a>
-          {% elif object.extension == '.mp4' %}
-            <div class="embed-responsive embed-responsive-16by9">
-              <video id="video-1_html_api" class="vjs-tech" preload="auto" autoplay >
-                <source src="{{ object.post_asset.url }}" type="video/mp4" class="embed-responsive">
-              </video>
-            </div>
-            <a href="{{ object.post_asset.url }}"  download class="text-dark"><h5>{{ object.post_asset }}</h5></a>
-          {% else %}
-            <a href="{{ object.post_asset.url }}"  download class="text-dark"><h5>{{ object.post_asset }}</h5></a>
-          {% endif %}
-        {% endif %}
-      </div>
-      <h2>{{ object.title }}</h2>
-      <p class="article-content">{{ object.post_content }}</p>
-    </div>
-  </article>
-{% endblock content %}
