--- blog/templates/form_of_post.html
@@ -1,18 +1,0 @@
-{% extends "blog/base.html" %}
-{% block content %}
-    <div class="content-section pt-2 mt-3 shadow-lg"  >
-          <form action="" method="POST" enctype="multipart/form-data">
-              {% csrf_token %}
-              <fieldset class="form-group">
-                    <legend class="border-bottom mb-4">
-                        Post
-                    </legend>
-              </fieldset>
-              <div class="form-group">
-                  
-                  <button class="btn btn-outline-primary" type="submit"><i class="fas fa-upload"></i> Upload</button>
-              </div>
-          </form>
-         
-      </div>
-{% endblock content %}
