--- blog/test.py
@@ -1,4 +1,0 @@
-
-from django.test import TestCase
-
-# Create your tests here.
