+++ config/urls.py
@@ -26,6 +26,5 @@ urlpatterns = [
     path('asset/<int:pk>/delete/',
          asset_views.delete_asset,
          name='delete_assets'),
-    path('', include('blog.urls')),
     #path('asset/list/', asset_views.list_assets, name='list_assets')
     ]
